Only the first `operator` in a chain gets special treatment
This makes it as easy to request "flatten" as it is to request "expansion". Previously we expanded if there was a line break anywhere, but that means you have to fully flatten a short expanded pipe chain to keep it flat.

Having special behavior for JUST the first operator is also more consistent with eventual behavior for calls and function definitions, where only the first argument/parameter will be treated specially.