Extend non-groupable comment check to comments in named arguments
If the trailing argument has any comments on the node, we bail on considering it for grouping and instead to the full expansion variant. We need to extend this check to the `name` and `value` nodes of an `RNamedArgument`, which JS doesn't have. This generally makes sense, as comments anywhere in here make it non-groupable automatically, but this is also important for idempotence purposes, because if it "looks" groupable here we can get unstable formatting.