Declare that we want git to check files out with unix line endings
This is critical for parser tests to run on Windows. Otherwise git will check them out with `\r\n` endings and you'll see a snapshot diff in the parser tests. It's what ruff and biome both do.

We will also need this file to explicitly declare `eof=crlf` on files when we want to add crlf specific tests