Split out `DefaultFilePatterns` with custom `matched_path_or_any_parents()` method
`matched_path_or_any_parents()` assumes you have provided a `root`, and it enforces this assumption by requiring that, after root stripping, `path` cannot contain root-like components, including a leading `/` on Unix and leading `C:/` on Windows. This is a reasonable restriction for ignore, but we've asserted that all of our patterns are root-agnostic by requiring `**/`, so we have to have a custom implementation that throws this check away.