Implement `exclude` and `default-exclude` TOML options (#213)
* Initial draft of `ignore` behavior

* Add `IncludePatterns` for "any R file" default inclusion rule

* Switch back to `exclude` to parallel `include`

* Use `"sourceLanguages": ["rust"]`

* Allow for specifying the test file URL

* Add tests for untitled and excluded files

* Add LSP test for respecting `exclude` and `default-excludes`

* Add `air.toml` that excludes parser/formatter snapshot files

* Remove test config

* Remove direct dependency on globset

* Update schema

* Split out `DefaultFilePatterns` with custom `matched_path_or_any_parents()` method

`matched_path_or_any_parents()` assumes you have provided a `root`, and it enforces this assumption by requiring that, after root stripping, `path` cannot contain root-like components, including a leading `/` on Unix and leading `C:/` on Windows. This is a reasonable restriction for ignore, but we've asserted that all of our patterns are root-agnostic by requiring `**/`, so we have to have a custom implementation that throws this check away.

* Fix fake file paths

* Add tests for `air format cpp11.R` and `air format renv`

* Add `matched_path_or_any_parents()` test for `DefaultFilePatterns`

* Sync `default_exclude` option name with setting name

* Polish `exclude` documentation a bit more