Skip semicolons as whitespace when not between two tokens (#240)
* Skip semicolons as whitespace when not between two tokens

In the `if between_two_token {` branch, we already implicitly consumed semicolons as `Whitespace` alongside actual whitespace. But before the first token and after the last token in a file, we were not consuming the semicolon "correctly". Ideally we get better support for this in the future, but for now this parses and gives us consisten behavior.

* CHANGELOG bullet