docs(neovim): add configuration guide for Neovim as LSP client
Introduce a new documentation file for configuring Neovim as a
Language Server Protocol (LSP) client using nvim-lspconfig and
conform.nvim. This guide provides step-by-step instructions for
setting up Air as an LSP client and enabling automatic formatting
on save. Update the main editors documentation to include a link
to the new Neovim guide, expanding the list of supported editors.