Implement Autobracing (#340)
* First draft of if statement autobracing

* Correctly recurse into `consequence` when handling non-enclosed comments

* Handle nested if statements in `consequence`

* Add failing test

* Don't pass pre-computed `braced_expression` value through to nested ifs in `consequence`

They should get to compute their own value, because the user could have validly written a one liner if statement inside a braced if to begin with, and we want to be consistent with that.

* Add one more test seen in data.table

* More test cases

* Another test case

* Another test

* Remove handling of comments not enclosed by the if statement

This causes too many issues with idempotence and verbatim handling, even though it would be quite nice

* Prefer generating leading comments

* Fix a typo buglet

* Only allow one liner if statements in specific contexts

* Add symmetrical support for function parameters, to mirror call arguments

* Refine end of line comments before `else`

* Remove context awareness from one liner computation

It seems like it is actually hard to get this right, as there are many other places where one liners feel "pretty good" and expanding them out causes more churn than it is worth.

* Add additional tests from trying it on dplyr

* Implement `SyntaxPosition` for `Value` vs `Effect` positioning

And use in if statement handling as a big improvement on context awareness!

* Tweak outdated comments

* Autobrace for loops, while loops, repeat loops, and function definitions (#344)

* Add autobracing for `for_statement`

* Add autobracing for `repeat_statement`

* Add autobracing for `while_statement`

* Add autobracing for `function_definition`

* In `for_statement`, push onto the `body` again

* In `while_statement`, push onto the `body` again

* In `repeat_statement`, push onto the `body` again

* Add documentation on autobracing

* CHANGELOG bullet

* Mention synergy with persistent line breaks

* Declare that any `RFunctionDefinition` can benefit from argument grouping

Allowing you to put a line break here `map(xs, function(x)<here>x + 1)`, which then causes autobracing to kick in, and now gives you the "middle variant" that you probably want of

```
map(xs, function(x) {
  x + 1
})
```

* Rework if statement section with value vs effect position

* Tweak CHANGELOG

* Portability is for multiline if statements

* Consistency with if statements

* Make effect/value comparison clearer

* Tweak a comment